+++ .babelrc.js
@@ -1,4 +1,14 @@
 module.exports = {
-	presets: ['@deloitte-digital-au/babel-preset-app-react'],
-	exclude: 'node_modules/**',
+	presets: [
+		'@babel/preset-env',
+		'@deloitte-digital-au/babel-preset-app-react',
+		'@babel/preset-typescript',
+	],
+	// Enables absolute imports
+	plugins: [
+		[
+			'module-resolver',
+			{ root: ['src'], extensions: ['.js', '.jsx', '.ts', '.tsx'] },
+		],
+	],
 };