+++ rollup.prod.js
@@ -1,3 +1,3 @@
 import generate from './rollup.common.js';
 
-export default generate(false);
+export default generate();