+++ src/index.ts
@@ -1,14 +1,11 @@
 import BreakpointContainer, {
 	AppBreakpoint,
-} from './components/BreakpointContainer';
-import BrowserContainer from './components/BrowserContainer';
-import {
-	withBreakpointContainer,
-	withBrowserContainer,
-} from './components/HOCs';
-import Breakpoint from './components/Breakpoint';
-import { bp, bpc } from './data/css-in-js';
-import { BREAKPOINTS, getBpUpperLimit, resolveBp } from './data/breakpoints';
+} from 'components/BreakpointContainer';
+import BrowserContainer from 'components/BrowserContainer';
+import { withBreakpointContainer, withBrowserContainer } from 'components/HOCs';
+import Breakpoint from 'components/Breakpoint';
+import { bpc } from 'data/css-in-js';
+import { BREAKPOINTS, getBpUpperLimit, resolveBp } from 'data/breakpoints';
 
 export {
 	// ---------------------